Partially apply #68
- Move flags to flags.go
- Move git-related code to git_main.go

We can do a proper refactoring later.
File separation works for now.
